fix(bridge): cast ShattersStatus.code to u32 for MSVC compatibility
MSVC defaults C enums to signed int, so bindgen generates
ShattersStatus.code as i32 on Windows but u32 on GCC/Clang.
Cast explicitly so BridgeError::Sdk { code: u32 } compiles on both.

+++ src-tauri/bridge/src/error.rs
@@ -37,7 +37,7 @@ pub(crate) fn check_status(mut status: ffi::ShattersStatus) -> BridgeResult<()> 
     };
 
     Err(BridgeError::Sdk {
-        code: status.code,
+        code: status.code as u32,
         message,
     })
 }
